Grouped finctionality into Game object
All game functionality is now encapsulated within the game object, which
has a method `step()` to do a single iteration of the game loop. This is
a good step towards creating an `Env` for the RL process. But for now
it helps clean-up the code.

Sidenote: Removed the `print()` statements for the player velocities.